aaa: The SSO token expiration date should be returned as Long
The SSO token expiration needs to returned as
a long value and should not be cast to int.

The token expiration is the Authorization
record expiration date returned from the
ldap server and should not be confused with
engine session expiration date.

Bug-Url: https://bugzilla.redhat.com/1391643

+++ backend/manager/modules/enginesso/src/main/java/org/ovirt/engine/core/sso/utils/SsoUtils.java
@@ -400,7 +400,7 @@ public class SsoUtils {
         ssoSession.setUserId(getUserId(principalRecord));
         try {
             ssoSession.setValidTo(validTo == null ?
-                    Integer.MAX_VALUE : (int) new SimpleDateFormat("yyyyMMddHHmmssZ").parse(validTo).getTime());
+                    Long.MAX_VALUE : new SimpleDateFormat("yyyyMMddHHmmssZ").parse(validTo).getTime());
         } catch (Exception ex) {
             log.error("Unable to parse Auth Record valid_to value: {}", ex.getMessage());
             log.debug("Exception", ex);
